+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `autoFocus` to `StyleguideInput`.
+
 ## [4.2.2] - 2019-09-20
 
 ## [4.2.1] - 2019-09-20